Update expectations to include url port numbers when Origin Isolation enabled.
When Origin Isolation is enabled (via kOriginKeyedProcessesByDefault),
the site_urls for SiteInstances will include port numbers when non-
standard ports are in use (which is the case with the embedded test
server used in browser tests). This CL updates the
FencedFrameTaskBrowserTests to include the port when Origin Isolation is
enabled.

Bug: 40259221
Change-Id: I371f03ca8f7843067582669f2667a0d1ebfab586
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5660619
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321125}