[FedCM] Increase spacing at bottom of multi IDP account chooser
Current spacing of 4 is too small so there is a jag visible at the
bottom left corner of the dialog. 7 is the smallest number where the jag
appears to not be visible, so we go with 8 since our margins are usually
even values.

Fixed: 350010042
Change-Id: I8f3f81079d436781fbc69e5df5083d950ff9bf01
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5663858
Commit-Queue: Nicolás Peña <npm@chromium.org>
Auto-Submit: Nicolás Peña <npm@chromium.org>
Reviewed-by: Zachary Tan <tanzachary@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321145}