CertProv: Add SignatureAlgorithm to ProofOfPossessionInstruction proto
Currently, there is only a single signature implemented and Client devices assume that this signature algorithm will be used. This behavior can cause problems when new signature algorithms are added in the future for which the client should perform different hashing and signing operations. By adding the `SignatureAlgorithm` enum to the `ProofOfPossessionInstruction` the Client is explicitly made aware of which signature algorithm the server-side code expects it to use, giving it the possibility to fail if the server-side intends to use a signature algorithm which the Client does not support.

Bug: b:348346219
Change-Id: Id180a6968e5c01cb9b79d4454d4e90c876dc6191
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5662304
Auto-Submit: Gregor Schwarz <gschwarz@google.com>
Reviewed-by: Michael Ershov <miersh@google.com>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Commit-Queue: Sergey Poromov <poromov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321158}