Rename enable_cxx GN variable to enable_rust_cxx
The enable_cxx name can be explicitly set in args.gn files, but it looks
strange, given how much of the codebase is in C++. Rename it to
enable_rust_cxx to give it some context. Leave enable_cxx as an alias
for now until downstream projects can migrate to the new name.

Bug: 40286066
Change-Id: Id727edbc2148e1bcd2b5a0d9450f6200640a15d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5665726
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321180}