[Fixit] Update tab strip to use Google Sans Text
Add a new utility class / method to apply attributes like text size,
color and font extracted from a TextAppearance style to a TextPaint
object. Use this method in TitleBitmapFactory to set these attributes to
the strip tab / group title TextPaint objects.

Since the font family in the TextAppearance style used for this is
already dictated by the FF #android-google-sans-text, this update need
not be explicitly flag-guarded.

This CL also partially fixes a linked issue with misalignment of the tab
group title text y position that surfaced during the font migration [2],
code in the cc layer needs to be updated to fix the persisting
smaller misalignment [4].

[1] With FF disabled (before crbug/339459839 fix):
https://screenshot.googleplex.com/4cpvChLkaYzDbLA
[2] With FF disabled (after crbug/339459839 fix): https://screenshot.googleplex.com/79WzhrLX8VyXsbB
[3] With FF enabled: https://screenshot.googleplex.com/Aujnjmiysf8xWoX
[4] crbug/339459839 TODO:
https://screenshot.googleplex.com/7pzHab9gguEuht5.png

updates in the way we update the TextPaint appearance using alternate
standard Android APIs, and in the calculation of the group text height.
Unit tests for the helper method introduced for the former have been
added for coverage.

Low-Coverage-Reason: TRIVIAL_CHANGE TitleBitmapFactory contains trivial
Bug: 348382514, 339459839
Change-Id: I40e9ffdaa6d3ab10f626f63fcb831e93ad230159
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5648507
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Aishwarya Rajesh <aishwaryarj@google.com>
Cr-Commit-Position: refs/heads/main@{#1321185}