Add installer target for linux-builder-perf builder
We want to generate symbol-stripped binaries for bisect archives
generated by linux-builder-perf. This CL adds an installer target to
linux-builder-perf and syncs the target chromedriver as
chromedriver_group as the official builder.

In addition to that, this CL also allows configuring the pinpoint
builders with different additional_compile_targets, as the targets that
generate stripped binaries might vary for perf builders.

Bug: 344643694
Change-Id: Icf3e1b9a80eff80b6dfb49ed42172a48d7fb5bbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5668198
Commit-Queue: Kuan Huang <kuanhuang@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321186}