[TabResumption] Fix crash in UrlImageProvider.
It seems the crash was caused the cached UrlImageProvider instance
isn't set to null after it is destroyed. In this CL, we check if the
instance has been destroyed before using it to stop the crash.

Bug: 350021945
Change-Id: I7d9c6dd1a8ccadec824a41a668003100b159fd52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5663859
Commit-Queue: Xi Han <hanxi@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Samuel Huang <huangs@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321192}