snap-groups: Fix re-snap with opposite & different snap ratio
This CL fixes the issue when re-snapping a window to the opposite side
with a different snap ratio. The underlying root cause is, when the
new snap event is received in `OnPreWindowStateTypeChange()`, we
break the group and send a nested bounds event that force updates the
event requested snap ratio. This adds a workaround that updates the
divider position with the new snap ratio first so the bounds can be
calculated correctly.

Demo: https://buganizer.corp.google.com/issues/349951979#comment5

Test: ash_unittests
Bug: b/349951979
Change-Id: Ica412a7ad842e8b8c9c20aa62d8188834b201578
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5668251
Reviewed-by: Michele Fan <michelefan@chromium.org>
Commit-Queue: Sophie Wen <sophiewen@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321201}