gpu: win: Move SCANOUT flag on CompoundImageBacking from gpu to shmem
This change moves the SCANOUT flag on Windows for NV12 SharedImages from
being on the gpu backing of a CompoundImageBacking to being on the
shared memory backing. This allows us to better distinguish on the
D3DImageBacking side which SCANOUT allocations are which, letting us
correctly identify the upcoming DComp texture backings.

Bug: 40150597
Change-Id: I18c973afbe498acbf0025746e194b245d768fa23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5662236
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Michael Tang <tangm@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1321203}