[remoting] Allow ChromeOS capturer to be started after SelectSource().
When multi-stream is enabled, the capturers are created per display
(using SelectSource() for each one), before they are started.

This CL fixes a crash and re-enables the multi-stream capability on
ChromeOS, by allowing SelectSource() to be called before Start().

The single-stream case did not crash, even though SelectSource() was
called first, because it was passed kFullDesktopScreenId which is not
an Ash Display ID and so it was ignored.

Bug: 335873769
Change-Id: Ie4e5ab07812db0ed15623d300adb4ade562dc0bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5664350
Commit-Queue: Lambros Lambrou <lambroslambrou@chromium.org>
Reviewed-by: Joe Downing <joedow@chromium.org>
Commit-Queue: Joe Downing <joedow@chromium.org>
Auto-Submit: Lambros Lambrou <lambroslambrou@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321209}