[omnibox][ml] Populate `search_suggest_relevance` for all Search matches
This CL updates the client-side code that populates the
`search_suggest_relevance` scoring signal for both query and
navigational Search suggestions. This addresses a shortcoming of the
original implementation, which only populated this signal for query
Search suggestions. Furthermore, this CL slightly modifies the criteria
for populating this signal to better align with the semantics stated
in omnibox_event.proto.

Bug: 335461869
Change-Id: I0c2c832ac6f222f9283e31009e3c3e8236f9cc93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5664307
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Commit-Queue: Khalid Peer <khalidpeer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321222}