viz: win: Explicitly mark SharedImage backings for DXGI swap chain usage
This change makes it so we explicitly mark SharedImage backings for the
root surface(s) on Windows to be DXGI swap chain or DComp surface.
Previously we needed to add the SI usage SCANOUT_DCOMP_SURFACE to
distinguish between a DXGI swap chain and DComp surface for SCANOUT
images. In the future, we will support DComp textures, which is another
kind of SCANOUT allocation, so we will need another flag to
differentiate between the tree.

Bug: 40244060
Change-Id: I3cf29caf9c3394a46fc2974442f498f8463d3682
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5662237
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Michael Tang <tangm@microsoft.com>
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1321262}