Fixed E2E mocking in ManualFillingControllerTest
Fixes the E2E mocking by returning true for #isEdgeToEdgeActive() as opposed to setting a non-zero value for the E2E bottom inset. Also declared the E2E supplier in order to be able to set the E2EController to null to fully disable E2E, which will be important as the #isEdgeToEdgeActive() method will be soon be removed.

Also flips the order of args in the asserts to match expected/actual.

Bug: 349670095
Change-Id: I617f164c3960f4b5976d157ca431a9040102d1f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5664105
Reviewed-by: Friedrich Horschig <fhorschig@chromium.org>
Commit-Queue: Charles Hager <clhager@google.com>
Cr-Commit-Position: refs/heads/main@{#1321263}