[Mac] Ellipsize long options rather than wrapping (to hidden) words that don't fit.
This changes the <select> popups on Mac so that they ellipsize
excessively large options that don't fit in the popup.  Prior to this
change the behavior was to attempt to wrap such options; however, the
wrapped text was not displayed.  In the case of an <optgroup> (which
causes indentation) the old behavior could mean that an option that was
a single large unbreakable unit could end up showing nothing at all.

Fixed: 348675991
Change-Id: I41f46048291f228c390f105b851f23055a0833d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5665600
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321282}