Remove unnecessary TTF call.
With crrev/c/5464824, a new call to TryToShowTouchToFill was introduced
that directly reacted to responses on OnGetSingleFieldSuggestions.

With the recent restructuring of `OnAskForValuesToFillImpl`,
`TryToShowTouchToFill` is now called before any
SingleFieldSuggestions, which makes a subsequent call in
OnGetSingleFieldSuggestionsCallback unnecessary.


Bug: 324557560
Change-Id: If7b6d5de9989ce0504cfd2b0e963e0095a944a03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5660448
Reviewed-by: Jan Keitel <jkeitel@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Norge Vizcay <vizcay@google.com>
Cr-Commit-Position: refs/heads/main@{#1321583}