Use wrapped context for drag and drop
We recently ran a Canary experiment for WebViewSeparateResourceContext.
Instead of crashes reducing as expected, they greatly increased. All of
the stack traces pointed to an issue with Drag and Drop. The experiment
affects how resources are loaded and when enabled, the Drag and Drop
code was unable to load its resources from the container view's context.
This CL replaces the Context used with one from the associated
WindowAndroid instead which can access Chromium resources.

Bug: 332746920
Change-Id: I34b46a4942325c19e5826ac1f4e9e4b8aad18675
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5650926
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Alex Mitra <alexmitra@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321621}