graphite: Fix YUVA promise image context leak
Graphite PromiseTextureFromYUVA has separate imageContext and
planeContexts. The ImageReleaseProc which is where Chromium cleans up
the FulfillForPlane contexts only gets called for the imageContext which
Chromium sets to null so we leak the planeContexts.

Fix that by making the imageContext be a heap array that holds the
FulfillForPlanes and the planeContexts are just pointers into the array.
The ImageReleaseProc deletes the heap array so there's no leak.

Bug: 350084287
Change-Id: Ibffebd1925e3883e98e7c66851ccb8a9d4fbdc51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5666599
Auto-Submit: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321654}