[DBSC] Ensure that BoundSessionParams::site is always canonical
Bound session storage wasn't cleared properly because a
non-canonicalized URL was stored in prefs. Once the URL was read from
the storage, it was converted to GURL which canonicalized in-memory
copy of the URL but not the copy in prefs.

The fix is essentially a one-line change in
`bound_session_registration_fetcher_impl.cc` that builds a GURL before
populating BoundSessionParams::site.

This would not fix the existing sessions though, so this CL also adds
the code that canonicalizes site URLs in storage. In addition, the
storage code is updated to remove invalid entries from prefs on stratup,
when the bound session service is initialized.

This CL also adds an extra condition
to `bound_session_credentials::AreParamsValid()` to verify that the site
is always canonicalized to prevent this error from reproducing in the
future.

Fixed: 349411334
Change-Id: Ifcedd000a718d50e3f1e6e3071a705a7e06c7c1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5660565
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Monica Basta <msalama@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321686}