[Omnibox] Remove fallback + left align text for simple card
The simple card shouldn't show a decoration when the icon would be
generic; instead, the content should be wider with a leadin matching the
action chips (16dp).

This is mostly accomplished by supporting a hidden decoration in
SuggestionLayout; hiding the decoration results in a minimum offset
being used instead of the decoration width.

Bug: 326368632
Change-Id: I8ff7df09414d388f21fdb384914d7cb19f3acdd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5666464
Reviewed-by: Tomasz Wiszkowski <ender@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Patrick Noland <pnoland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321712}