[Fixit] Update TextInputLayout text styles part 3
Some of our TextInputLayouts have nested AutoCompleteTextViews, which
ignore ?attr/editTextStyle that we define in the TextInputLayout theme
overlay. This CL adds ?attr/autoCompleteTextViewStyle to the theme
overlay to make sure the text has the correct font regardless of the
type of EditText we use.

Bug: 348400515
Change-Id: I5e1603f7a7f8db23059bd512071304aa9c2f4a3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5665177
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Sinan Sahin <sinansahin@google.com>
Cr-Commit-Position: refs/heads/main@{#1321737}