Fix test expectations for recently-imported transform animation tests.
transform-interpolation-007.html is not actually flaky.  The expectation
file is correct, and the log claiming it was flaky was a log from the
WPT importer run that was generating the expectation file.

transform-non-invertible-no-transition.html is flaky on all platforms,
not just Mac and Windows.

Bug: 349883135
Change-Id: Iaa3085feba45fa2427f7fa4783e8cccc6aa637f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5670884
Commit-Queue: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Auto-Submit: David Baron <dbaron@chromium.org>
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1321781}