Add simple bottom chin MVC implementation
go/e2e-bottom-chin-design-doc

This is a simple first implementation for the bottom chin, which seeks to implement a new bottom browser controls component to act as a scrollable bottom navbar. This first pass sets up the file structure for the C++ and MVC for the component, as well as some rudimentary logic.

Bug: 345383907
Change-Id: I3a55ee309dc7007a7a461be126faebcd96a2fd10
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5601926
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Charles Hager <clhager@google.com>
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321804}