Stop sending UncaughtJsError UKM event on every exception
This event is very spammy for certain clients and has been impacting
the UKM.Entries.Dropped alerts. We are disabling this event for now
until we can put in place a cap on the number of times this event can be
sent per client over some time period.

Bug: b/338241225
Change-Id: I67a4d3632f248e1bd626154d3387b4ea0063f07e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5665871
Commit-Queue: Ryan Kalla <ryankalla@google.com>
Reviewed-by: Jonathan Njeunje <njeunje@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321809}