[iOS] Update Contextual Panel bottom inset based on toolbar height
Before this, the panel did not know the height of the toolbar, so the
bottom of the content got stuck behind the toolbar. This CL fixes that
by alerting the panel content of the bottom toolbar's expanded height.
It does not attempt to on-the-fly change the content inset when the
toolbar's height changes due to fullscreen.

Bug: 343727945
Change-Id: I6ba1696b42193aae1c0079e7c0cdf907a39eb875
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5660865
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Aliona Dangla <alionadangla@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Nicolas MacBeth <nicolasmacbeth@google.com>
Cr-Commit-Position: refs/heads/main@{#1321820}