[Autofill] Fix FormFieldData::value semantics for select elements
Usually, FormFieldData::value matches the IDL attribute, but
sometimes Autofill determines the value of select elements
differently:

If ExtractOptions::kOptionText is set, FormFieldData::value is
populated with the selected option's text instead of its value.
However, ExtractOptions::kOptionText is not set consistently.

This CL fixes this inconsistency by moving the
ExtractOption::kOptionText logic to the browser process.

More precisely:

(1) This CL introduces a getter
  FormFieldData::selected_option()
that matches the behaviour of ExtractOption::kOptionText: it returns
the selected option's text if there is one, and otherwise the select's
value. In corner cases, the behaviour differs from
ExtractOption::kOptionText:
- FormFieldData::selected_option() operates on truncated strings,
  so it may match a wrong option if two values have the same
  1024-character-long prefix;
- if there are >512 <option> elements, none of them is extracted
  and hence FormFieldData::selected_option() is nullopt;
- FormFieldData::options() are populated with the aria-label as
  fallback if the option's text is empty, which however is likely
  an extremely rare case.

(2) This CL replaces the certain references to FormFieldData::value
with FormFieldData::selected_option(). Those references are the ones
where, before this CL, the field's
- FormFieldData::form_control_type may be FormControlType::kSelect* and
- FormFieldData::value was usually set according to the
  ExtractOption::kOptionText logic. ("Usually" because
  ExtractOption::kOptionText wasn't set consistently.)

(3) This CL removes ExtractOption::kOptionText, so that
FormFieldData::value is now always populated with the form control's
"value" IDL attribute.

More details on this change and the audit of the existing
FormFieldData::value references are in go/autofill-select-value.

Bug: 349901783, 40227496, 40232021
Change-Id: I627d772bbc47b65b425e434b8065126991d7f927
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5666510
Commit-Queue: Christoph Schwering <schwering@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1321877}