Handle encrypted archive entries in UpdateArchiveAnalyzerResultsWithFile
Currently these are handled in the archive-specific analyzers. This
means that each analyzer has to remember to set is_encrypted
correctly. We had missed updating the seven_zip_analyzer.cc in
https://crrev.com/c/5484410.

To be more robust for the future, handle this in the common helper
function called on each entry.

Bug: 349860783
Change-Id: I267833a4992b2a5b2ec88cea2ce52c508dfb8789
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5671282
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321880}