snap-groups: Fix re-snap to opposite side via layout menu
This CL fixes the issue where re-snapping to the opposite side with
the layout menu would auto-group every other time. The fix is to *not*
end overview early if the mouse event is in the layout menu, such that
we remain in partial overview.

Demo: https://buganizer.corp.google.com/issues/349892870#comment4

Test: added, verified it fails w/o fix in Patchset 1
Bug: b/349892870
Change-Id: I29d2799999b99242302a27037375dbf1ad4c23b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5670818
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Sophie Wen <sophiewen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321896}